requirements.txt
alembic==1.15.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
databricks-sdk==0.47.0
Deprecated==1.2.18
docker==7.1.0
exceptiongroup==1.2.2
fastapi==0.115.12
flask==3.1.0
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib-metadata==8.6.1
importlib-resources==6.5.2
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
liac-arff==2.5.0
Mako==1.3.9
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mlflow==2.21.0
mlflow-skinny==2.21.0
narwhals==1.32.0
numpy==2.0.2
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.6
pydantic-core==2.27.2
pydeck==0.9.1
plotly
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
rsa==4.9
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.39
sqlparse==0.5.3
starlette==0.46.1
streamlit==1.43.2
seaborn==0.13.2
tenacity==9.0.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
typing-extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
waitress==3.0.2
watchdog==6.0.0
werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
